requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
django-user-agents==0.4.0
django-ipware==6.0.0 